==> modeling/multi_output_chain_forecasting/code/Step00_Create_Folders.py <==
# =============================================================================
# library
# =============================================================================
import os

# =============================================================================
# load data
# =============================================================================
os.mkdir("output/step1")
os.mkdir("output/step2")
os.mkdir("output/step3")
os.mkdir("output/step4")
